requirements.txt
streamlit==1.38.0
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.2.0

